crew/pyproject.toml
[project]
name = "crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.98.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = []
